requirements.txt
dm_sonnet==1.23
graph_nets==1.0.2
mujoco_py.egg==info
baselines.egg==info
seaborn==0.9.0
cloudpickle==0.6.1
gym==0.10.9
networkx==2.2
tensorflow_probability==0.5.0
numpy_quaternion==2018.12.2.0.32.47
ipdb==0.11
anyrl.egg==info
matplotlib==3.0.2
pyperclip==1.7.0
mpi4py==3.0.0
numpy==1.15.4
scipy==1.1.0
tensorflow_gpu==1.12.0
opencv_python==3.4.5.20
setuptools==40.6.3
Pillow==6.1.0
anyrl==0.12.23
baselines==0.1.5
mujoco_py==2.0.2.2
pyautogui==0.9.45
quaternion==0.03.1
sonnet==0.1.6
tensorflow==1.14.0
tf==1.0.0
PyYAML==5.1.1
